requirements.txt
fastapi
uvicorn
python-dotenv
pillow
google-genai
python-multipart
pydantic
pydantic-settings
pydantic[email]
supabase
pytz